Add test cases for COFACTOR and generalize SW tests
